gen v2: Make generic TypeScript template work
