+++ gen/template_typescript_test.go
@@ -1,14 +1,14 @@
-package typescript
+package gen
 
 import (
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/webrpc/webrpc/gen"
 	"github.com/webrpc/webrpc/schema"
 )
 
-const input = `
+const tsInput = `
 {
   "webrpc": "v1",
   "name": "example",
@@ -160,14 +160,15 @@ const input = `
 `
 
 func TestGenTypescript(t *testing.T) {
-	s, err := schema.ParseSchemaJSON([]byte(input))
+	schema, err := schema.ParseSchemaJSON([]byte(tsInput))
 	assert.NoError(t, err)
 
-	g := &generator{}
+	tsTemplates := os.DirFS("./typescript")
 
-	o, err := g.Gen(s, gen.TargetOptions{})
-	assert.NoError(t, err)
-	_ = o
+	o, err := Generate(schema, tsTemplates, TargetOptions{})
+	if err != nil {
+		t.Fatal(err)
+	}
 
-	// t.Logf("%s", o)
+	t.Logf("%s", o)
 }
