+++ gen/typescript/client_helpers.ts.tmpl
