+++ gen/typescript/proto.gen.ts.tmpl
